+++ clients/admin-ui/src/features/data-discovery-and-detection/action-center/constants.ts
@@ -4,3 +4,15 @@ export enum DiscoveryStatusDisplayNames {
   EXEMPT = "Exempt",
   UNKNOWN = "Unknown",
 }
+
+export enum DiscoveredAssetsColumnKeys {
+  NAME = "name",
+  RESOURCE_TYPE = "resource_type",
+  SYSTEM = "system",
+  DATA_USES = "data_uses",
+  LOCATIONS = "locations",
+  DOMAIN = "domain",
+  PAGE = "page",
+  CONSENT_AGGREGATED = "consent_aggregated",
+  ACTIONS = "actions",
+}
